matz
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@836 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ st.c
@@ -546,14 +546,14 @@ strhash(string)
 
 static int
 numcmp(x, y)
-    int x, y;
+    long x, y;
 {
     return x != y;
 }
 
 static int
 numhash(n)
-    int n;
+    long n;
 {
     return n;
 }
